Updates InputStreamSpec -> InputStreamTest
